perf: improve performance of AutoMergeBlendShape
This pass was obtaining the blendshape buffer for each vertex, then processing those buffers.
However, in practice, most vertices share the same buffer; this pass ignored this, and redid
the same work thousands of times as a result.

This reduces processing time from ~370ms to ~22ms on my main avatar.