perf: avoid fetching mask texture data on change check callbacks
NDMF invokes change checks (Observe callbacks) periodically to check for changes
caused by code that does not properly register change events. While this is time-sliced,
if any single callback takes too long, it can cause poor editor performance.

In AAO, we were checking for changes in mask textures in such a callback; this check retrieved
the `GetPixels32` array, which is an expensive operation which could take dozens of ms in some cases
(particularly since the time-slicing was on a per-object basis, so if the same mask was used by
multiple objects, the overhead adds up).

While NDMF should probably change the time slicing to work on a per-listener basis, and/or add
special handling for `Texture2D`, on the AAO side we can more efficiently perform change checks by\
using unity's built-in image hash functions instead of comparing contents directly.