fix: complete graph to entry exit broke immediate transition
The issue is miminoko v2 hard mode is broken.
This is because miminoko v2 uses 'always true' condition to enter a state for single frame, but C2E/E will transform to A>A rather than A>B>A.
This commit fixes this by denying transforming C2E/E for possible immediate transition.
This may reduce the optimization target cases, but the main target of this optimization; C2E/E2BlendTree does not affected by this change